feat(boost): Change boost timer implementation ######################################################################## In order to reduce zigbee traffic, we throttle down how often we send updates to the valves. We only send an update every 10 messages from the valves.
The timer boost is also no longer tracked by ticks, but instead by a future date in which it
should stop. This works better for restarts as the restart could theoretically happen after the
boost time has ended. The change from ticks to date implied the lovelace card now has to
work with dates.