feat(boost): Fix boost date management
Properly uses dates to determine boost in the component and in the UI.

In the component, the boost_end date is used on restore to set the boost end timer.
If the difference between the boost_end date and "now" is negative, i.e. the boost
has ended, the difference is set to 0 so the boost_end timer is called immediately
and the boost ends.

In the card, the boost_end time is properly converted to a dayjs object so the same
comparison can be used. The difference is used to calculate the filled circle.